+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Unit test for `WebArchive.to_html()`.
 ### Changed
 * Don't delete the `srcset` attribute from `<img>`.
 * Embed style sheets in single-file mode using data URIs rather than `<style>`.